Use Singleton to reduce model load times
